Move DWARF symbolization benchmark into benches/symbolize
We initially organized benchmarks by source type (e.g., DWARF, Gsym,
etc.). However, by now we changed the overall organization of the crate
in such a way that different areas of functionality (symbolization,
inspection, normalization) are separated.
With this change we start organizing benchmarks similarly, by moving
the DWARF symbolization benchmark into benches/symbolize.
