+++ benches/dwarf.rs
@@ -2,33 +2,10 @@ use std::path::Path;
 
 use blazesym::inspect;
 use blazesym::inspect::Inspector;
-use blazesym::symbolize::Elf;
-use blazesym::symbolize::Source;
-use blazesym::symbolize::Symbolizer;
 
 use criterion::measurement::Measurement;
 use criterion::BenchmarkGroup;
 
-
-/// Symbolize an address, end-to-end, i.e., including all necessary setup.
-fn symbolize_end_to_end() {
-    let dwarf_vmlinux = Path::new(&env!("CARGO_MANIFEST_DIR"))
-        .join("data")
-        .join("vmlinux-5.17.12-100.fc34.x86_64");
-    let src = Source::Elf(Elf::new(dwarf_vmlinux));
-    let symbolizer = Symbolizer::new();
-
-    let results = symbolizer
-        .symbolize(&src, &[0xffffffff8110ecb0])
-        .unwrap()
-        .into_iter()
-        .flatten()
-        .collect::<Vec<_>>();
-    assert_eq!(results.len(), 1);
-
-    let result = results.first().unwrap();
-    assert_eq!(result.symbol, "abort_creds");
-}
 
 /// Lookup an address, end-to-end, i.e., including all necessary setup.
 fn lookup_end_to_end() {
@@ -57,9 +34,6 @@ where
     if cfg!(feature = "generate-bench-files") {
         group.bench_function(stringify!(dwarf::lookup_end_to_end), |b| {
             b.iter(lookup_end_to_end)
-        });
-        group.bench_function(stringify!(dwarf::symbolize_end_to_end), |b| {
-            b.iter(symbolize_end_to_end)
         });
     }
 }
