+++ benches/symbolize.rs
@@ -1,4 +1,7 @@
+use std::path::Path;
+
 use blazesym::c_api;
+use blazesym::symbolize::Elf;
 use blazesym::symbolize::Process;
 use blazesym::symbolize::Source;
 use blazesym::symbolize::Symbolizer;
@@ -25,6 +28,27 @@ fn symbolize_process() {
     assert_eq!(results.len(), addrs.len());
 }
 
+/// Symbolize an address in a DWARF file, end-to-end, i.e., including all
+/// necessary setup.
+fn symbolize_dwarf() {
+    let dwarf_vmlinux = Path::new(&env!("CARGO_MANIFEST_DIR"))
+        .join("data")
+        .join("vmlinux-5.17.12-100.fc34.x86_64");
+    let src = Source::Elf(Elf::new(dwarf_vmlinux));
+    let symbolizer = Symbolizer::new();
+
+    let results = symbolizer
+        .symbolize(&src, &[0xffffffff8110ecb0])
+        .unwrap()
+        .into_iter()
+        .flatten()
+        .collect::<Vec<_>>();
+    assert_eq!(results.len(), 1);
+
+    let result = results.first().unwrap();
+    assert_eq!(result.symbol, "abort_creds");
+}
+
 pub fn benchmark<M>(group: &mut BenchmarkGroup<'_, M>)
 where
     M: Measurement,
@@ -32,4 +56,9 @@ where
     group.bench_function(stringify!(symbolize::symbolize_process), |b| {
         b.iter(symbolize_process)
     });
+    if cfg!(feature = "generate-bench-files") {
+        group.bench_function(stringify!(symbolize::symbolize_dwarf), |b| {
+            b.iter(symbolize_dwarf)
+        });
+    }
 }
